lowering Mojolicious version for old Perl versions
Running the following command on the Mojolicious repo:

    git show 953ddc87219fa79e02f5884de3051088048c8a8d:lib/Mojolicious.pm | grep VERSION

returns 7.31

This is the commit [1] where the slurp/spurt functions have been removed
from Mojo::Util.

But earlier than that, in [2] the slurp/spurt methods were added to the Mojo::File
package:

    git show 0afb80bdc05b995b81a80a02af2c0ed52f3b4e3c:lib/Mojolicious.pm | grep VERSION

in version 7.15

We'll use 7.15 as a required Mojolicious version for the old Perl
versions.

[1] https://github.com/mojolicious/mojo/commit/953ddc87219fa79e02f5884de3051088048c8a8d
[2] https://github.com/mojolicious/mojo/commit/0afb80bdc05b995b81a80a02af2c0ed52f3b4e3c